Paddle
The paddle has now been added and the movement seems to work and is stopping on the screen edges, but not dynamically. So, if the screen is resized, it will stop at the border of the 500p screen. I will change the brick and paddle size/positioning to adjust dynamically for screen resizing before moving onto the ball.