Background Music
Changed it so the music plays in the background the whole time instead of start screen. Check this out to see if you like it or not. Super easy to change so we dont have to keep it if we dont want. We can also turn the volume down of the music too.